% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessing.R
\name{annotateGenes}
\alias{annotateGenes}
\title{annotateGenes  genes with reference to hg38}
\usage{
annotateGenes(mtx)
}
\arguments{
\item{mtx}{raw count matrix}
}
\value{
annotations of each genes in rows with chrom and positions.
}
\description{
annotateGenes  genes with reference to hg38
}
\examples{
test.count_mtx_annot <- annotateGenes(mat=matx)
}
