% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunc.R
\name{annoteBandOncoHeat}
\alias{annoteBandOncoHeat}
\title{annoteBandOncoHeat Annotate with chromosome bands the data frame with difference copy number alterations between subclones}
\usage{
annoteBandOncoHeat(mtx_annot, diffSub, nSub, organism = "human")
}
\arguments{
\item{mtx_annot}{Annotation matrix}

\item{diffSub}{Data frame with difference copy number alterations between subclones}

\item{nSub}{Number of subclones}

\item{organism}{Organism to be analysed (default = "human")}
}
\description{
annoteBandOncoHeat Annotate with chromosome bands the data frame with difference copy number alterations between subclones
}
