% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyTumor.R
\name{baseline.synthetic}
\alias{baseline.synthetic}
\title{estimate basline copy numbers by synthetic normal cell}
\usage{
baseline.synthetic(norm.mat = norm.mat, min.cells = 10, par_cores = 20)
}
\arguments{
\item{norm.mat}{smoothed data matrix; genes in rows; cell names in columns.}

\item{min.cells}{minimal number of cells per cluster.}

\item{par_cores}{number of cores for parallel computing.}
}
\value{
1) relative gene expression; 2) synthetic baseline profiles; 3) clustering results.
}
\description{
estimate basline copy numbers by synthetic normal cell
}
\examples{

count_mtx_relat <- test.relt$expr.relat
}
