% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyTumor.R
\name{classifyCluster}
\alias{classifyCluster}
\title{Classify the two clusters on the basis of confident normal cells}
\usage{
classifyCluster(hcc2, norm.cell.names)
}
\arguments{
\item{hcc2}{Two clusters from hierarchical clustering}

\item{norm.cell.names}{Vector of confident normal cells}
}
\value{

}
\description{
Classify the two clusters on the basis of confident normal cells
}
