% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyTumor.R
\name{classifyTumorCells}
\alias{classifyTumorCells}
\title{classifyTumorCells classify tumour and normal cells from the raw count matrix}
\usage{
classifyTumorCells(
  count_mtx,
  annot_mtx,
  sample = "",
  distance = "euclidean",
  par_cores = 20,
  ground_truth = NULL,
  norm.cell.names = NULL,
  SEGMENTATION_CLASS = TRUE
)
}
\arguments{
\item{count_mtx}{raw count matrix}

\item{annot_mtx}{matrix containing the annotations of the genes}

\item{sample}{sample name (optional)}

\item{distance}{distance used in hierarchical clustering (optional)}

\item{par_cores}{number of cores (optional)}

\item{norm.cell.names}{confident normal cells (optional)}

\item{SEGMENTATION_CLASS}{Boolean value to perform segmentation before classification (optional)}

\item{gr_truth}{ground truth of classification (optional)}
}
\value{

}
\description{
classifyTumorCells classify tumour and normal cells from the raw count matrix
}
\examples{

}
