% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegaMC.R
\name{getBreaksVegaMC}
\alias{getBreaksVegaMC}
\title{getBreaksVegaMC Get SCEVAN segmentation of the matrix.}
\usage{
getBreaksVegaMC(mtx, chr_vect, sample = "", beta_vega = 0.5)
}
\arguments{
\item{mtx}{count matrix}

\item{chr_vect}{Vector specifying for each gene the chromosome where it is located}

\item{sample}{sample name (optional)}

\item{beta_vega}{specifies beta parameter for segmentation, higher beta for more coarse-grained segmentation. (default 0.5)}
}
\value{
breakpoints
}
\description{
getBreaksVegaMC Get SCEVAN segmentation of the matrix.
}
