% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidentNormal.R
\name{getConfidentNormalCells}
\alias{getConfidentNormalCells}
\title{Get at most top 30 confident normal cells}
\usage{
getConfidentNormalCells(mtx, par_cores = 20)
}
\value{

}
\description{
Get at most top 30 confident normal cells
}
\examples{

}
