% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSampleComparisonClonalCN.R
\name{multiSampleComparisonClonalCN}
\alias{multiSampleComparisonClonalCN}
\title{multiSampleComparisonClonalCN Compare the clonal Copy Number of multiple samples.}
\usage{
multiSampleComparisonClonalCN(
  listCountMtx,
  listNormCells = NULL,
  analysisName = "all",
  organism = "human",
  par_cores = 20,
  plotTree = TRUE
)
}
\arguments{
\item{listCountMtx}{Named list of raw count matrix of samples}

\item{analysisName}{Name of the analysis (default "all")}

\item{organism}{Organism to be analysed (optional - "mouse" or "human" - default "human")}

\item{par_cores}{number of cores (default 20)}
}
\description{
multiSampleComparisonClonalCN Compare the clonal Copy Number of multiple samples.
}
