% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineCNA.R
\name{pipelineCNA}
\alias{pipelineCNA}
\title{Run pipeline runs the pipeline that classifies tumour and normal cells from the raw count matrix and looks for possible sub-clones in the tumour cell matrix}
\usage{
pipelineCNA(
  count_mtx,
  sample = "",
  par_cores = 20,
  gr_truth = NULL,
  SUBCLONES = TRUE
)
}
\arguments{
\item{count_mtx}{raw count matrix}

\item{sample}{sample name (optional)}

\item{par_cores}{number of cores (optional)}

\item{gr_truth}{ground truth of classification (optional)}

\item{SUBCLONES}{find subclones (optional)}
}
\value{

}
\description{
Run pipeline runs the pipeline that classifies tumour and normal cells from the raw count matrix and looks for possible sub-clones in the tumour cell matrix
}
\examples{
res_pip <- pipelineCNA(count_mtx, par_cores = 20, gr_truth = gr_truth, SUBCLONES = TRUE)
}
