% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSampleComparisonClonalCN.R
\name{plotAllClonalCN}
\alias{plotAllClonalCN}
\title{Title plotAllClonalCN}
\usage{
plotAllClonalCN(samples, name)
}
\arguments{
\item{samples}{Vector with sample names to be plotted}

\item{name}{Analysis name}
}
\description{
Title plotAllClonalCN
}
