% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSampleComparisonClonalCN.R
\name{plotAllSubclonalCN}
\alias{plotAllSubclonalCN}
\title{plotAllSubclonalCN Plot the copy number of each subclone of a sample.}
\usage{
plotAllSubclonalCN(sample, pathOutput = "./output/")
}
\arguments{
\item{sample}{Name of the sample.}

\item{pathOutput}{Path to the output folder containing the output of pipelineCNA.}
}
\description{
plotAllSubclonalCN Plot the copy number of each subclone of a sample.
}
