% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotCNA_withAnnotCells}
\alias{plotCNA_withAnnotCells}
\title{plotCNA_withAnnotCells allows generating a heatmap of the copy number profile of each cell, adding cell annotations as tracks on the heatmap.}
\usage{
plotCNA_withAnnotCells(
  SampleName,
  metadata,
  COLUMNS_TO_PLOT,
  outputPATH = "./output/",
  SUBCLONE = FALSE,
  hcc = NULL,
  plotNAME = "heatmap_with_annotation.png",
  par_cores = 20
)
}
\arguments{
\item{SampleName}{Sample name used in pipelineCNA}

\item{metadata}{data.frame cells (rownames) and annotations (columns)}

\item{COLUMNS_TO_PLOT}{columns of metadata to be added as tracks in the heatmap}

\item{outputPATH}{output folder of pipelineCNA (optional)}

\item{SUBCLONE}{Boolean value TRUE if you are interested in CNA matrix of sublclone and FALSE if you are interested in CNA matrix of all cells.}

\item{hcc}{if you have previously computed clustering for the heatmap (optional - default 0.5)}

\item{plotNAME}{name file to save the figure (optional)}

\item{par_cores}{number of cores used for clustering (optional - default 20)}
}
\description{
plotCNA_withAnnotCells allows generating a heatmap of the copy number profile of each cell, adding cell annotations as tracks on the heatmap.
}
\examples{
plotCNA_withAnnotCells(SampleName, metadata, c("CellType","Tissue","Cluster"))
}
