% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessing.R
\name{preprocessingMtx}
\alias{preprocessingMtx}
\title{preprocessingMtx}
\usage{
preprocessingMtx(
  count_mtx,
  sample,
  ngenes_chr = 5,
  perc_genes = 0.1,
  par_cores = 20,
  findConfident = TRUE,
  AdditionalGeneSets = NULL,
  SCEVANsignatures = TRUE,
  organism = "human"
)
}
\arguments{
\item{count_mtx}{raw count matrix}

\item{ngenes_chr}{minimum number of genes per chromosome (optional)}

\item{perc_genes}{percentage of cells in which each gene is to be expressed
(optional)}

\item{par_cores}{number of cores (optional)}

\item{findConfident}{Boolean value to search for normal cells (default TRUE)}

\item{AdditionalGeneSets}{List of additional signatures to be used to search
for normal cells (optional)}

\item{SCEVANsignatures}{Boolean value TRUE to use internal SCEVAN signatures
for normal cells or FALSE to use only signatures specified in
AdditionalGeneSets (default TRUE)}

\item{SMOOTH}{Boolean value to perform smoothing (optional)}
}
\value{
count_mtx_smooth processed and smoothed matrix
count_mtx_annot annotated matrix
}
\description{
Pre-processing steps: Cells with less than 200 genes and the genes expressed
in less than \verb{1\%} of cells are removed. Genes are annotated and sorted
according to genomic coordinates. Highly confident normal cells are sought in
the matrix. Genes involved in the cell cycle pathway are removed.
Log-Freeman–Tukey transformation to stabilize variance and a polynomial
dynamic linear modeling (DLM) to smooth out the outliers.
}
