% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineCNA.R
\docType{package}
\name{SCEVAN}
\alias{SCEVAN-package}
\alias{SCEVAN}
\title{SCEVAN: R package that automatically classifies the cells in the scRNA data by segregating non-malignant cells of tumor microenviroment from the malignant cells. It also infers the copy number profile of malignant cells, identifies subclonal structures and analyses the specific and shared alterations of each subpopulation.}
\description{
The SCEVAN package
}
\section{SCEVAN functions}{

The SCEVAN functions ...
}

