% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortData.R
\name{sortData}
\alias{sortData}
\title{Sorts a dataset file by the genomic position of the probes.}
\usage{
sortData(dataset, output_file_name = "")
}
\arguments{
\item{dataset}{Dataset file.}

\item{output_file_name}{Name of the file in which sorted data are stored.}
}
\value{
the input matrix ordered by the genomic position of the probes.
}
\description{
Sorts a dataset file by the genomic position of the probes.
}
\note{
This function allows to sort a dataset by the genomic position. The
input file must have the chromosome and the position in column two and three
respectively. This format follows the standard output of PennCNV. An example
of file can be found in inst/example folder.
}
\references{
Morganella S., and Ceccarelli M. VegaMC: a R/bioconductor package for fast
downstream analysis of large array comparative genomic hybridization
datasets. Bioinformatics, 28(19):2512-4 (2012).
}
