% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessing.R
\name{annotateGenes}
\alias{annotateGenes}
\title{annotateGenes Annotate genes with genomic coordinates with reference to hg38 using Ensembl based annotation package}
\usage{
annotateGenes(mtx)
}
\arguments{
\item{mtx}{Count matrix with genes on row names (Ensemble or Symbol)}
}
\value{
Annotated matrix
}
\description{
annotateGenes Annotate genes with genomic coordinates with reference to hg38 using Ensembl based annotation package
}
\examples{
count_mtx_annot <- annotateGenes(count_mtx)
}
