% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyTumor.R
\name{classifyCluster}
\alias{classifyCluster}
\title{classifyCluster Classify the two major clusters of CNA matrix on the basis of confident normal cells}
\usage{
classifyCluster(hcc2, norm_cell_names)
}
\arguments{
\item{hcc2}{Two clusters from hierarchical clustering}

\item{norm_cell_names}{Vector of confident normal cells}
}
\value{
classification of tumor and normal cells
}
\description{
classifyCluster Classify the two major clusters of CNA matrix on the basis of confident normal cells
}
