% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyTumor.R
\name{classifyTumorCells}
\alias{classifyTumorCells}
\title{func}
\usage{
classifyTumorCells(
  count_mtx_smooth,
  count_mtx_proc,
  count_mtx_annot,
  sam.name,
  distance = "euclidean",
  par_cores = 20,
  ground_truth = NULL,
  norm.cell.names = NULL,
  UP.DR = 0.1,
  ngene.chr = 5,
  WRITE = FALSE,
  SEGMENTATION_CLASS = TRUE
)
}
\value{

}
\description{
func
}
\examples{

}
