% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyTumor.R
\name{computeCNAmtx}
\alias{computeCNAmtx}
\title{computeCNAmtx computed the CNA matrix using the break points obtained from segmentation}
\usage{
computeCNAmtx(count_mtx, breaks, par_cores = 20, segmAlt)
}
\arguments{
\item{count_mtx}{count matrix}

\item{par_cores}{number of cores for parallel computing (optional)}

\item{breaksbreak}{points obtained from segmentation}
}
\value{
CNA matrix
}
\description{
computeCNAmtx computed the CNA matrix using the break points obtained from segmentation
}
\examples{

}
