% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessing.R
\name{preprocessingMtx}
\alias{preprocessingMtx}
\title{func}
\usage{
preprocessingMtx(
  count_mtx,
  ngenes_chr = 5,
  perc_genes = 0.1,
  par_cores = 20,
  SMOOTH = TRUE
)
}
\value{

}
\description{
func
}
\examples{


}
