% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessing.R
\name{preprocessingMtx}
\alias{preprocessingMtx}
\title{func}
\usage{
preprocessingMtx(
  count_mtx,
  ngenes_chr = 5,
  perc_genes = 0.1,
  par_cores = 20,
  SMOOTH = TRUE
)
}
\arguments{
\item{count_mtx}{raw count matrix}

\item{ngenes_chr}{minimum number of genes per chromosome (optional)}

\item{perc_genes}{percentage of cells in which each gene is to be expressed (optional)}

\item{par_cores}{number of cores (optional)}

\item{SMOOTH}{Boolean value to perform smoothing (optional)}
}
\value{

}
\description{
func
}
\examples{

}
