% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyTumor.R
\name{removeSyntheticBaseline}
\alias{removeSyntheticBaseline}
\title{removeSyntheticBaseline Removes a synthetic baseline from a tumour pure matrix}
\usage{
removeSyntheticBaseline(count_mtx, par_cores = 20)
}
\arguments{
\item{count_mtx}{count matrix}

\item{par_cores}{number of cores for parallel computing.}
}
\value{
relative matrix
}
\description{
removeSyntheticBaseline Removes a synthetic baseline from a tumour pure matrix
}
\examples{

}
