% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunc.R
\name{subclonesTumorCells}
\alias{subclonesTumorCells}
\title{Title}
\usage{
subclonesTumorCells(tum_cells, CNAmat, sample)
}
\arguments{
\item{tum_cells}{}

\item{CNAmat}{}

\item{sample}{}
}
\value{

}
\description{
Title
}
