% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidentNormal.R
\name{top30classification}
\alias{top30classification}
\title{Get at most top 30 confident normal cells}
\usage{
top30classification(
  NES,
  pValue,
  FDR,
  pval_filter,
  fdr_filter,
  pval_cutoff,
  nes_cutoff,
  nNES
)
}
\arguments{
\item{NES}{}

\item{pValue}{}

\item{FDR}{}

\item{pval_filter}{}

\item{fdr_filter}{}

\item{pval_cutoff}{}

\item{nes_cutoff}{}

\item{nNES}{}
}
\value{

}
\description{
Get at most top 30 confident normal cells
}
